Move routing-code around
